Moved setupTableName() back into concrete class
git-svn-id: https://cida-svn.er.usgs.gov/repos/dev/usgs/dcpt/trunk@16011 622dc1fb-2224-0410-ab72-dc85a3975e1a

+++ src/test/java/gov/usgs/cida/ncetl/spec/IngestControlSpecTest.java
@@ -1,5 +1,6 @@
 package gov.usgs.cida.ncetl.spec;
 
+import gov.usgs.webservices.jdbc.spec.Spec;
 import java.io.IOException;
 import java.sql.SQLException;
 import javax.naming.NamingException;
@@ -21,7 +22,7 @@ import static org.hamcrest.Matchers.*;
  */
 public class IngestControlSpecTest {
 
-    IngestControlSpec test = new IngestControlSpec();
+    Spec test;
     String DB_LOCATION = FileHelper.getTempDirectory() +  "test_delete_me" + File.separator;
     Connection connection;
     String DB_FULL_LOCATION = DB_LOCATION + "test.db";
@@ -49,7 +50,11 @@ public class IngestControlSpecTest {
         if (connection != null && !connection.isClosed()) {
             connection.close();
         }
-        FileUtils.forceDelete(new File(DB_LOCATION));
+        File dbLoc = new File(DB_LOCATION);
+        if (dbLoc.exists()) {
+            FileUtils.forceDelete(new File(DB_LOCATION));
+        }
+        
     }
 
     @Test
